refactor(components): better direcotory structure for components
Components used specifically grouped together.
