--- src/components/Navbar/_lib/consts/nav-items.tsx
@@ -1,17 +1,0 @@
-export const navItems = [
-  {
-    text: 'About',
-    fullpath: '/',
-    path: '',
-  },
-  {
-    text: 'Writing',
-    fullpath: '/writing',
-    path: 'writing',
-  },
-  {
-    text: 'Misc',
-    fullpath: '/misc',
-    path: 'misc',
-  },
-];
